server/pyproject.toml
[tool.aerich]
tortoise_orm = "config.TORTOISE"
location = "./migrations"
src_folder = "./."

[tool.pdm]
[[tool.pdm.source]]
url = "https://pypi.org/simple"
verify_ssl = true
name = "pypi"
[tool.pdm.dev-dependencies]
dev = [
    "black[d]>=22.8.0",
    "pylama",
    "aiosqlite",
    "mypy>=0.981",
]

[project]
name = ""
version = ""
description = ""
authors = [
    {name = "apaudel96", email = "apaudel96@ymail.com"},
]
dependencies = [
    "uvicorn[standard]",
    "passlib[argon2]",
    "aerich>=0.7.1",
    "tortoise-orm[asyncpg]",
    "arrow",
    "fastapi",
    "python-multipart",
    "nanoid>=2.0.0",
    "python-jose[cryptography]>=3.3.0",
    "python-dotenv>=0.21.0",
]
requires-python = ">=3.10"
license = {text = "MIT"}
[project.optional-dependencies]

[build-system]
requires = ["pdm-pep517>=1.0.0"]
build-backend = "pdm.pep517.api"
